--- app/models/ping_pong.rb
@@ -1,2 +1,0 @@
-module PingPong
-end
