+++ app/services/betting/spread.rb
