+++ app/services/elo_adjustment.rb
