+++ app/services/ping_pong/finalization.rb
