+++ app/services/ping_pong/service.rb
