+++ app/services/ping_pong/table_controls.rb
