+++ app/services/slack_notification.rb
