+++ app/services/stats/streak_adjustment.rb
