+++ spec/services/ping_pong/finalization_spec.rb
