+++ spec/services/ping_pong/rewind_spec.rb
