+++ spec/services/ping_pong/service_spec.rb
