+++ spec/services/ping_pong/table_controls_spec.rb
