+++ spec/services/stats/streak_adjustment_spec.rb
